requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
passlib[bcrypt,argon2]==1.7.4
pyotp==2.9.0
requests
qrcode
Pillow
cryptography